apps/desktop/requirements.txt
PyQt6>=6.7.0
qasync>=0.27.1
google-genai>=1.0.0
google-cloud-aiplatform>=1.60.0
google-adk>=0.1.0
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.21.0
google-cloud-secret-manager>=2.20.0
google-cloud-logging>=3.10.0
google-cloud-storage>=2.16.0
google-auth>=2.29.0
pyautogui>=0.9.54
playwright>=1.44.0
mss>=9.0.2
sounddevice>=0.4.7
numpy>=1.26.0
Pillow>=10.3.0
pynput>=1.7.7
pvporcupine>=3.0.0
websockets>=12.0
firebase-admin>=6.5.0
pyinstaller>=6.6.0
python-dotenv>=1.0.1
